pyproject.toml
[tool.poetry]
name = "electronic-shop"
version = "0.1.0"
description = ""
authors = ["hoolywant <sir.saltickow@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
pep8 = "^1.7.1"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
coreapi = "^2.3.3"
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
